move some stack utility functions to common
