+++ src_common/starkDisplayUtils.h
@@ -1,0 +1,6 @@
+#ifdef HAVE_STARKWARE
+
+void stark_sign_display_master_account();
+void stark_sign_display_condition_fact();
+
+#endif
